Moved rest of uitl.cpp
