+++ src/cuda_internal.h
@@ -1,4 +1,5 @@
 #include "internal.h"
+#include "log.h"
 
 struct CUDAThreadState: ThreadState{
     
@@ -31,6 +32,32 @@ struct CUDAThreadState: ThreadState{
 
     /// Perform a synchronous copy operation
     void jitc_memcpy(void *dst, const void *src, size_t size) override;
+    
+    /// Perform an assynchronous copy operation
+    void jitc_memcpy_async(void *dst, const void *src, size_t size) override;
+    
+    /// Replicate individual input elements to larger blocks
+    void jitc_block_copy(enum VarType type, const void *in, void *out,
+                                 uint32_t size, uint32_t block_size) override;
+
+    /// Sum over elements within blocks
+    void jitc_block_sum(enum VarType type, const void *in, void *out,
+                                uint32_t size, uint32_t block_size) override;
+    
+    /// Asynchronously update a single element in memory
+    void jitc_poke(void *dst, const void *src, uint32_t size) override;
+
+    void jitc_aggregate(void *dst, AggregationEntry *agg,
+                                uint32_t size) override;
+
+    // Enqueue a function to be run on the host once backend computation is done
+    void jitc_enqueue_host_func(void (*callback)(void *),
+                                        void *payload) override;
+    
+    /// LLVM: reduce a variable that was previously expanded due to dr.ReduceOp.Expand
+    void jitc_reduce_expanded(VarType vt, ReduceOp op, void *data, uint32_t exp, uint32_t size) override{
+        jitc_raise("jitc_reduce_expanded(): unsupported by CUDAThreadState!");
+    }
 
     ~CUDAThreadState(){}
 };
